update long value , scoreUI

+++ Assets/Scripts/SquareData.cs
@@ -6,11 +6,11 @@ public class SquareData
 {
     public Utils.Cell cell;
     public int id;
-    public int value;
+    public long value;
 
     public Vector2 Position => Utils.GridToPos(cell.Row, cell.Column);
 
-    public SquareData(Utils.Cell cell, int id, int value)
+    public SquareData(Utils.Cell cell, int id, long value)
     {
         this.cell = cell;
         this.id = id;
