+++ Assets/Scripts/Utils.cs
@@ -14,7 +14,7 @@ public static class Utils
         return new Cell((int)((pos.x + 5) / 2), (int)((4- pos.y) / 2));
     }
 
-    public static int GetExponent(int value)
+    public static int GetExponent(long value)
     {
         return (int)(Mathf.Log(value) / Mathf.Log(2));
     }
